+++ sdk/storage/storage-file/test/retrypolicy.spec.ts
@@ -1,13 +1,14 @@
 import * as assert from "assert";
+import * as dotenv from "dotenv";
 
 import { RestError, StorageURL } from "../src";
 import { Aborter } from "../src/Aborter";
-import { ShareURL } from "../src/ShareURL";
 import { Pipeline } from "../src/Pipeline";
+import { ShareURL } from "../src/ShareURL";
 import { getBSU } from "./utils";
 import { InjectorPolicyFactory } from "./utils/InjectorPolicyFactory";
 import { record } from "./utils/recorder";
-import * as dotenv from "dotenv";
+
 dotenv.config({ path: "../.env" });
 
 describe("RetryPolicy", () => {
@@ -51,6 +52,35 @@ describe("RetryPolicy", () => {
 
     const result = await shareURL.getProperties(Aborter.none);
     assert.deepEqual(result.metadata, metadata);
+  });
+
+  it("Retry Policy should abort when abort event trigger during retry interval", async () => {
+    let injectCounter = 0;
+    const injector = new InjectorPolicyFactory(() => {
+      if (injectCounter < 2) {
+        injectCounter++;
+        return new RestError("Server Internal Error", "ServerInternalError", 500);
+      }
+    });
+
+    const factories = shareURL.pipeline.factories.slice(); // clone factories array
+    factories.push(injector);
+    const pipeline = new Pipeline(factories);
+    const injectShareURL = shareURL.withPipeline(pipeline);
+
+    const metadata = {
+      key0: "val0",
+      keya: "vala",
+      keyb: "valb"
+    };
+
+    let hasError = false;
+    try {
+      await injectShareURL.setMetadata(Aborter.timeout(2 * 1000), metadata);
+    } catch (err) {
+      hasError = true;
+    }
+    assert.ok(hasError);
   });
 
   it("Retry Policy should fail when requests always fail with 500", async () => {