+++ sdk/storage/storage-queue/src/policies/RetryPolicy.ts
@@ -1,18 +1,18 @@
 import {
+  AbortSignalLike,
   BaseRequestPolicy,
-  delay,
   HttpOperationResponse,
   HttpPipelineLogLevel,
   RequestPolicy,
   RequestPolicyFactory,
   RequestPolicyOptions,
   RestError,
-  WebResource
+  WebResource,
 } from "@azure/ms-rest-js";
 
 import { IRetryOptions } from "../RetryPolicyFactory";
 import { URLConstants } from "../utils/constants";
-import { setURLHost, setURLParameter } from "../utils/utils.common";
+import { delay, setURLHost, setURLParameter } from "../utils/utils.common";
 
 /**
  * A factory method used to generated a RetryPolicy factory.
@@ -53,8 +53,10 @@ const DEFAULT_RETRY_OPTIONS: IRetryOptions = {
   retryDelayInMs: 4 * 1000,
   retryPolicyType: RetryPolicyType.EXPONENTIAL,
   secondaryHost: "",
-  tryTimeoutInMs: 30 * 1000 //https://docs.microsoft.com/en-us/rest/api/storageservices/setting-timeouts-for-queue-service-operations
+  tryTimeoutInMs: 30 * 1000 // https://docs.microsoft.com/en-us/rest/api/storageservices/setting-timeouts-for-queue-service-operations
 };
+
+const RETRY_ABORT_ERROR = new RestError("The request was aborted", RestError.REQUEST_ABORTED_ERROR);
 
 /**
  * Retry policy with exponential retry and linear retry implemented.
@@ -195,7 +197,7 @@ export class RetryPolicy extends BaseRequestPolicy {
       }
     }
 
-    await this.delay(isPrimaryRetry, attempt);
+    await this.delay(isPrimaryRetry, attempt, request.abortSignal);
     return await this.attemptSendRequest(request, secondaryHas404, ++attempt);
   }
 
@@ -296,10 +298,11 @@ export class RetryPolicy extends BaseRequestPolicy {
    * @private
    * @param {boolean} isPrimaryRetry
    * @param {number} attempt
+   * @param {AbortSignalLike} [abortSignal]
    * @returns
    * @memberof RetryPolicy
    */
-  private async delay(isPrimaryRetry: boolean, attempt: number) {
+  private async delay(isPrimaryRetry: boolean, attempt: number, abortSignal?: AbortSignalLike) {
     let delayTimeInMs: number = 0;
 
     if (isPrimaryRetry) {
@@ -319,6 +322,6 @@ export class RetryPolicy extends BaseRequestPolicy {
     }
 
     this.logf(HttpPipelineLogLevel.INFO, `RetryPolicy: Delay for ${delayTimeInMs}ms`);
-    return delay(delayTimeInMs);
+    return delay(delayTimeInMs, abortSignal, RETRY_ABORT_ERROR);
   }
 }